fix(notes): resolve dark mode PDF rendering inconsistency between /notes and /notes/:uuid pages
- Create shared pdf-generator utility to eliminate code duplication
- Fix /notes page PDF generation to properly detect and apply dark mode
- Ensure consistent PDF formatting across all note download locations
- Add txt download support to Notes.svelte for feature parity
